Deprecate usage of XML Spring schemas
Move BadClassPath test to annotation

Related to #141.
