+++ src/test/java/fr/pilato/spring/elasticsearch/it/annotation/transport/mappingconvention/AppConfig.java
@@ -20,19 +20,13 @@
 package fr.pilato.spring.elasticsearch.it.annotation.transport.mappingconvention;
 
 import fr.pilato.spring.elasticsearch.ElasticsearchTransportClientFactoryBean;
-import org.elasticsearch.client.Client;
-import org.springframework.context.annotation.Bean;
+import fr.pilato.spring.elasticsearch.it.annotation.transport.TransportAppConfig;
 import org.springframework.context.annotation.Configuration;
 
-import java.util.Properties;
-
 @Configuration
-public class AppConfig {
-	@Bean
-	public Client esClient(Properties esProperties) throws Exception {
-		ElasticsearchTransportClientFactoryBean factory = new ElasticsearchTransportClientFactoryBean();
-		factory.setProperties(esProperties);
-		factory.afterPropertiesSet();
-		return factory.getObject();
+public class AppConfig extends TransportAppConfig {
+	@Override
+	protected void enrichFactory(ElasticsearchTransportClientFactoryBean factory) {
+
 	}
 }
